refactored SimpleDemo to class
