+++ etc/config.example
@@ -3,7 +3,6 @@
 
 # App common stuff
 APP_NAME="GenAlgo"
-APP_SRC_DIR="./src"
 
 ## Logging
 APP_VAR_DIR="./var/"
