+++ etc/config_requirements.inc.php
@@ -3,7 +3,6 @@
 
 /* @var \Dotenv\Dotenv $config */
 
-$config->required('APP_SRC_DIR');
 $config->required('APP_CACHE_DIR');
 $config->required('APP_LOG_DIR');
 $config->required('APP_LOG_LEVEL');
