+++ src/GenAlgo/SolutionException.php
@@ -1,3 +1,5 @@
 <?php
 
+namespace GenAlgo;
+
 class SolutionException extends \Exception {}
