--- src/SimpleDemo.php
@@ -1,371 +1,0 @@
-<?php
-/*
-TODO: MANN CHECK:
-7b9ebe0412690fcbb70f239be54b059a556cb948 100 0.7 0.09609375 9 42 10000 2016-08-30T08:23:56+02:00 10677/12000 5000 TIMEOUT
-
-ORG: 101101011000110100010011001111000011,001111000000010110011101001010111001,101001011000100101000101100000100101,100001010000110100100110100000010111,101101010110000010000101000100000000,000000101100101011010011010010000111,101000111010000110110001001010110100,100001010101000100101011100000110110,110101000110000001011000110000011000,000101010011100100100110011001000100,011101100001110000000110010000101001,000110111100011110110010110110110111,100010101001001111000001011001110011,110100000111001110000000010001100001,101110100100011110110001010101011100,100001011010001011010110101100110101,110010011101001010001101101011000111,101000111101110000001010001010001000,010011011101001010000100110110100100,011010000111101101100011101010011100,100101011010001000101101001000000000,110000101000011001111000011010010010,101010011101000100010111100111000000,110010001001101001000000011101100011,011010010011011110000110000100001101,101001101001110000110100110001000110,010111010100010111001101000010010011,000100101010010010000110100000100000,000001000000110100000111100111011011,110110110001010100100001101001110000,101010001001000010011011101000000110,001110001000001110001100010010001100,101100111100100000111001101010011100,110001010101110000100000100000101010,010111011010101100100101011001100000,010010101000001110001100000100110001,101111011010100111000010000110110100,000101010111110010110110100010011001,000000010001000001101011100010010110,011010111001100010001001010001000111,011101010101101101111011010001011000,101100000100011100010101100000011011,011010100111110000110101011110110000,001000010100101010001010000101100100,110110111001011110011010101100101011,001110101101000100111001100000101100,000010101001000011001011010010010110,000010111101010010101010000001000110,101100101000100101010101100001101000,001100011011001000010111110000010110,101001100001001001111100001011001001,110111011101010110100010110101100111,010100000000000001000010101101110100,010001010110101000011100110001000110,101101100100011001010011011010100000,100010100110001000010111001100011100,010111010101101000111010001011011100,000111000010100010011000110000010000,001010001010001100100111101001011001,010010001011001000001010011110101101,010011001101001000001011010110000110,110101111000110110100011101011010110,001110101011110100000101101000100110,011010100010010100010001010101000001,001010011010100110000011010010000000,011101001101000010001010101101111010,001101001101100110101010110000101011,110101110001000110101010110001100101,001010101101001001000100001001001101,110000100110100001101011100000010111,010000001001000111010010001100011101,110111010101010100010010010000110110,100101011011100001001101000111000110,110001110111010110111000000111010111,010000101001001100101001100101111010,101111000000010001110110000100101011,000100111001011100010010000101110000,100110001101001111010001110000100100,100010111011010010011010010111010100,110000000110100100101001010110101011,100010110100100001111101100010101101,100110010001000000111101110010001001,100000001000110111011001010110011011,000100000111110010010101001100111100,001110111001001001110101010001111001,010001100100110100000101011111010100,001001011101000001100000100001011001,000010011101110111001011100000010100,000001011100101010100101000010010101,010100110101101001101011100101110100,101000011001010100010100010000000011,000110010100011010101010010101100111,101001110010001011000101100010001100,010101000101100100000110010101101000,100110111001110111000101010001010010,000110101000100001110001101110010000,001100111001000110010000011001000001,110110010111100001010100010001000011,100110011000101110100101011001001100,100000011000100001001011001101100111
-FIN: 1,11,10,00,11,00,11,10,11,11,10,00,1,10,1,1,00,1,01,1,1,1,1,10,01,11,10,1,1,1,1,10,10,11,1,10,1,1,00,11,1,1,1,10,1,1,1,10,10,01,1,00,10,10,00,1,1,1,00,1,10,1,1,1,00,10,00,1,00,10,11,1,10,1,11,10,10,11,1,01,11,00,10,10,1,1,10,11,00,11,11,01,10,01,1,1,1,10,00,11,11,1,10,1,11,10,00,1,11,1,00,1,1,10,1,1,1,11,1,10,1,1,1,00,11,10,1,1,1,11,11,1,1,1,01,10,10,1,1,10,00,1,11,1,11,1,10,1,1,1,11,1,1,1,1,00,11,1,10,10,10,10,11,11,01,1,1,10,1,11,1,11,1,10,1,1,00,1,1,1,1,11,11,1,11,00,11,1,1,11,1,00,10,00,00,10,1,10,1,1,1,1,10,1,1,01,00,00,11,1,10,10,11,1,1,1,00,1,01,10,1,10,11,1,1,1,1,00,10,1,1,11,00,1,11,1,10,00,10,1,1,1,10,1,00,11,11,00,1,00,1,1,1,11,1,10,1,1,11,00,1,1,1,11,10,1,1,10,11,1,11,1,00,11,1,1,1,1,10,10,1,10,01,10,1,10,1,1,01,00,10,10,1,1,1,1,1,11,1,1,1,00,10,1,11,11,1,1,1,1,1,11,00,1,00,1,00,00,11,1,11,00,01,10,01,01,00,1,00,1,01,10,10,01,10,11,10,00,1,11,11,1,11,1,10,1,1,00,01,11,1,1,1,1,1,1,1,1,1,1,10,1,1,11,1,1,1,1,10,1,1,10,01,1,10,11,10,01,1,11,1,1,10,00,1,11,11,10,1,00,01,00,1,1,1,1,1,1,1,11,1,1,10,11,10,1,1,1,11,10,01,10,11,10,10,01,1,1,11,1,01,00,01,01,00,1,11,1,00,10,00,1,1,1,10,00,10,1,1,10,10,1,1,1,10,1,1,1,11,01,11,10,1,1,1,10,1,11,1,1,1,1,1,11,00,1,1,1,00,11,00,1,11,1,1,1,1,1,1,1,11,1,1,10,1,11,1,1,1,10,11,1,1,00,1,00,11,00,1,1,1,1,1,1,01,01,01,1,1,1,1,00,1,00,1,10,1,1,1,1,1,01,1,10,10,00,1,11,00,1,00,10,1,00,11,01,11,10,10,1,1,1,10,10,10,1,11,1,1,1,1,10,01,10,10,1,1,1,1,11,10,1,10,1,1,10,10,00,10,10,00,00,01,1,10,1,1,1,1,1,10,1,11,00,1,11,00,1,10,1,11,1,00,1,1,1,00,00,10,1,1,1,10,10,00,01,01,00,1,1,1,1,10,1,1,00,01,1,11,1,1,01,01,1,1,1,10,1,10,00,1,1,01,10,11,10,1,10,1,1,10,11,10,1,10,01,1,00,1,1,1,10,1,10,1,10,00,10,00,00,01,00,1,11,1,1,00,11,1,1,00,11,01,1,1,1,1,1,11,01,1,10,1,1,1,1,1,1,10,11,1,11,1,11,10,01,1,01,00,1,1,10,1,11,1,10,10,00,1,00,1,10,1,10,1,1,1,11,1,11,1,1,10,1,01,1,1,1,1,1,1,1,1,1,10,1,1,1,1,10,01,1,00,1,00,10,10,00,10,11,01,01,10,01,00,11,01,1,1,1,00,1,1,1,1,1,11,10,00,1,1,00,10,1,1,1,1,10,01,11,00,1,1,11,1,00,11,11,00,1,1,1,00,11,11,10,10,10,11,1,1,11,1,01,1,1,10,00,10,1,1,00,10,10,00,11,11,01,1,10,1,1,1,1,1,10,1,11,11,10,1,10,1,1,1,1,1,1,1,1,11,1,10,1,1,1,1,11,11,10,10,11,10,11,1,00,1,10,1,1,01,11,1,10,1,10,1,10,1,1,11,00,1,1,00,00,1,1,10,00,01,01,1,1,00,11,10,00,1,01,1,1,1,1,00,00,11,11,1,1,10,1,1,1,10,00,11,1,1,1,10,1,1,11,10,11,1,11,00,11,00,1,1,1,10,1,1,1,10,10,1,00,11,10,1,1,1,1,1,1,1,1,11,1,1,1,1,11,1,1,10,11,10,1,1,1,1,1,1,10,11,1,1,1,1,10,00,01,10,1,1,1,1,11,1,11,01,10,1,1,01,01,11,1,1,1,10,11,1,1,10,10,1,01,11,1,1,01,1,10,1,11,10,11,10,01,10,1,1,00,1,11,00,1,1,11,00,00,1,1,1,1,1,1,11,1,11,01,00,1,1,1,1,1,1,1,1,00,10,1,1,1,1,10,10,00,1,11,1,11,00,1,10,1,00,1,1,1,11,10,1,1,11,10,10,01,01,1,1,1,1,1,1,10,00,1,11,10,11,1,1,1,1,10,1,11,00,11,1,10,1,10,01,01,1,00,10,1,1,1,10,10,01,1,1,1,1,11,00,1,1,11,1,1,10,1,11,1,00,10,1,1,11,00,01,10,11,11,10,1,00,10,11,11,10,01,1,1,10,1,1,10,11,1,10,11,10,1,11,1,1,00,01,1,1,1,10,11,1,11,11,1,10,1,00,10,1,11,1,11,1,1,1,1,1,10,1,1,1,10,11,1,1,1,10,1,01,00,11,11,00,1,10,10,11,10,11,1,10,1,1,1,1,11,00,1,00,00,10,11,1,1,1,1,1,00,1,00,11,1,11,1,1,1,11,01,10,00,11,01,10,1,1,10,11,1,11,10,11,1,1,1,01,01,10,11,00,10,1,10,10,01,10,1,1,10,10,1,11,11,10,10,11,10,01,10,00,11,01,1,1,1,1,10,11,1,1,1,00,00,00,1,1,11,1,1,01,00,01,00,00,10,11,00,1,1,1,1,00,10,01,11,1,10,1,00,11,1,1,1,1,1,11,1,00,11,1,11,1,1,11,00,1,10,11,1,10,1,00,1,11,1,11,1,01,11,1,1,00,1,10,1,1,10,11,10,1,11,1,1,1,1,1,10,00,10,11,10,11,00,11,1,10,1,10,1,01,10,11,1,1,11,01,11,10,1,00,1,10,1,10,1,1,1,1,11,1,01,01,10,1,1,10,1,11,1,1,10,01,1,11,00,11,1,00,1,1,10,1,1,1,1,00,00,1,11,1,1,1,11,1,00,1,00,1,10,1,00,11,11,10,11,00,1,11,10,11,10,10,1,01,11,11,01,10,11,1,00,01,1,11,1,1,11,1,1,1,1,1,10,10,01,11,00,1,10,1,00,11,10,1,11,11,1,1,1,1,1,00,1,1,1,11,1,1,1,1,1,1,1,10,1,00,1,10,1,10,1,10,1,10,10,1,1,1,1,1,1,1,1,1,1,1,1,1,10,1,1,10,1,1,10,1,1,10,1,1,1,1,10,11,00,10,1,11,10,1,10,00,01,11,01,01,00,1,1,00,1,1,1,10,11,1,1,00,10,1,10,1,11,00,01,1,10,1,1,1,10,1,10,11,10,1,1,1,00,1,1,10,1,1,10,00,1,00,00,11,01,10,1,1,11,1,11,01,10,1,1,00,10,1,1,10,1,1,1,11,10,1,1,1,10,1,11,1,1,1,01,11,01,00,1,11,1,11,1,1,1,1,11,00,10,1,10,1,1,1,10,01,10,1,00,1,11,1,11,1,00,1,1,1,1,10,00,1,00,1,10,10,1,10,1,1,01,00,10,11,10,11,00,10,1,00,11,00,11,10,10,00,00,10,1,11,11,11,1,1,10,1,10,1,11,11,01,00,1,1,01,10,1,10,1,10,10,1,10,10,1,10,00,01,00,1,1,11,00,01,1,1,1,1,1,00,10,10,11,1,01,1,01,01,1,1,1,11,1,10,1,1,1,1,1,01,11,1,01,00,1,10,11,1,1,1,11,1,11,1,11,01,10,00,1,00,1,11,1,1,1,1,1,01,10,01,10,10,10,1,1,11,00,1,1,1,1,1,1,10,00,11,1,1,1,1,1,1,11,00,10,11,11,1,1,10,10,11,1,10,00,1,10,1,10,1,10,1,10,11,1,1,10,01,1,11,1,10,1,01,00,00,01,1,1,1,11,1,00,1,1,1,10,1,1,1,1,1,11,00,11,1,1,1,11,1,11,1,1,10,01,10,1,11,1,10,1,1,00,1,10,1,1,1,11,1,00,00,10,1,00,1,10,1,10,1,1,11,10,1,10,11,11,10,1,00,1,1,1,1,10,1,1,1,1,01,10,1,11,1,10,00,00,10,10,10,10,1,11,1,1,10,11,10,11,1,1,1,00,1,11,1,10,1,1,1,00,01,00,10,11,10,1,1,10,1,1,00,00,1,1,1,1,1,1,1,1,00,1,11,10,01,1,11,1,1,1,1,1,00,11,00,1,1,1,11,1,11,01,1,11,1,1,1,1,1,00,1,00,1,01,10,11,1,1,1,1,1,10,1,1,1,11,10,1,1,11,10,1,01,11,00,11,1,1,1,10,1,1,01,11,01,10,1,1,00,10,1,1,01,00,1,00,1,1,1,00,1,1,1,1,1,10,1,11,10,01,1,10,1,1,1,1,1,1,11,11,10,01,01,1,11,1,10,11,10,1,00,1,1,1,1,1,1,1,1,1,00,01,10,01,1,10,01,10,1,10,1,1,1,00,10,1,10,1,00,01,10,10,11,1,1,10,1,1,1,11,1,1,10,11,1,00,01,11,00,10,1,1,1,1,1,1,1,1,01,00,1,10,01,1,10,1,10,10,00,1,1,11,1,01,11,00,1,11,1,1,1,1,10,10,1,1,10,1,11,1,1,11,11,1,10,1,1,1,11,1,10,1,1,01,00,10,1,1,1,1,1,10,1,1,1,01,10,1,10,11,10,1,00,10,1,00,00,10,00,1,1,11,00,1,10,00,11,1,00,1,10,11,00,10,1,00,00,11,01,00,1,00,00,10,11,11,10,1,1,1,1,1,1,1,1,1,1,11,10,11,00,01,10,00,01,10,1,01,00,1,10,00,11,1,11,1,1,1,01,00,11,1,1,1,10,1,11,10,1,1,10,00,10,11,10,1,1
-
-2016-08-30T10:42:27+02:00 1472538236.3416 1472546547.3498 8311.0082371235
-*/
-
-use GenAlgo\ConfigurationValues;
-
-require_once (__DIR__ . '/SimpleCode.php');
-require_once (__DIR__ . '/SolutionException.php');
-require_once (__DIR__ . '/SelectionException.php');
-// todo:mann test 24k with 0.09609375 AND 0.09375 each in one process
-
-const CODE = [
-    '0000' => '0',
-    '0001' => '1',
-    '0010' => '2',
-    '0011' => '3',
-    '0100' => '4',
-    '0101' => '5',
-    '0110' => '6',
-    '0111' => '7',
-    '1000' => '8',
-    '1001' => '9',
-    '1010' => '+',
-    '1011' => '-',
-    '1100' => '*',
-    '1101' => '/',
-];
-const CHROMOSOME_LENGTH = 9;
-
-# interesting MUTATION_RATES
-#const MUTATION_RATE     = 0.09375;
-#const MUTATION_RATE     = 0.09609375;
-
-/**
- * @param $searchedValue
- * @param ConfigurationValues $c
- * @return array
- */
-function simpleDemo($searchedValue, ConfigurationValues $c) {
-  // todo: mann: store intial population of long running
-  // todo: mann: test optimal parameters
-  // todo: mann: are optimal parameters generic or problem related?
-  $code = new SimpleCode(CODE);
-  $population = initPopulation($code, $c->PopulationSize(), CHROMOSOME_LENGTH);
-  $intialPopulation = $population;
-  $counter = 1;
-
-  while (true) {
-    try {
-      $new = generateNewPopulation($population, $code, $c, $searchedValue);
-    } catch(\SolutionException $e) {
-      return [$counter, $e->getMessage()];
-    } catch(\SelectionException $e) {
-      return [
-        $counter,
-        $e->getMessage(),
-        implode(',', $intialPopulation),
-        implode(',', $population)
-      ];
-    }
-
-    $counter++;
-    # echo "New population found: $counter" . PHP_EOL;
-    $population = $new;
-
-    if ($counter >= $c->MaxPopulations()) {
-      return [
-        $counter,
-        'TIMEOUT',
-        implode(',', $intialPopulation),
-        implode(',', $population)
-      ];
-    }
-  }
-}
-
-/**
- * @param SimpleCode $code
- * @param $size
- * @param $length
- * @return array
- */
-function initPopulation(SimpleCode $code, $size, $length) {
-  $population = [];
-
-  for ($i = 0; $i < $size; $i++) {
-    $population[] = generateNewSpez($code, $length);
-  }
-
-  return $population;
-}
-
-/**
- * @param SimpleCode $code
- * @param $length
- * @return string
- */
-function generateNewSpez(SimpleCode $code, $length) {
-  $newSpez = '';
-
-  for ($i = 0; $i < $length; $i++) {
-      $newSpez .= $code->randomGen();
-  }
-
-  return $newSpez;
-}
-
-/**
- * @param array $population
- * @param SimpleCode $code
- * @param ConfigurationValues $c
- * @param $searched
- * @return array
- */
-function generateNewPopulation(array $population, SimpleCode $code, ConfigurationValues $c, $searched) {
-  $newPopulation = [];
-  $fitness = testPoputlation($population, $code, $searched);
-  $populationFitness = calcFitnessSum($fitness) / count($population);
-  # echo "avg population fitness: $populationFitness" . PHP_EOL;
-
-  while(count($population) > count($newPopulation)) {
-    list($spez1, $spez2) = selectPair($fitness, $c);
-    # echo "Selected >> ";
-    # echo implode(' ', $spez1) . " - ";
-    # echo implode(' ', $spez2) . PHP_EOL;
-
-    list($newSpez1,$newSpez2) = sex($spez1['chromosome'], $spez2['chromosome'], $c);
-    # echo "Children >> $newSpez1 - $newSpez2" . PHP_EOL;
-
-    if (!empty($newSpez1)) {
-      $newPopulation[] = $newSpez1;
-    }
-
-    if (!empty($newSpez2)) {
-      $newPopulation[] = $newSpez2;
-    }
-  }
-
-  return $newPopulation;
-}
-
-/**
- * @param array $population
- * @param SimpleCode $code
- * @param $searched
- * @return array
- */
-function testPoputlation(array $population, SimpleCode $code, $searched) {
-  $fitnessAbsolute = [];
-
-  foreach ($population as $spez) {
-    # echo "Testing spez $spez >> ";
-    $decoded = $code->decode($spez);
-    # echo "$decoded >> ";
-
-    $fitness = calculateFittness($searched, $decoded);
-    # echo "Fittness: $fitness ";
-    # echo "... done" . PHP_EOL;
-
-    $fitnessAbsolute[] = [
-      'chromosome' => $spez,
-      'fitnessAbsolute' => $fitness
-    ];
-  }
-
-  $fitnessRelative = calcRelativeFitness($fitnessAbsolute);
-  usort($fitnessRelative, function($a, $b) {
-    // usort casts result to integer, so we have to return a big
-    // integer which is big enough to meet our precision requirements
-    // @see http://us2.php.net/manual/en/function.usort.php
-    return
-      $a['fitnessRelative'] * 1000000000000000 -
-      $b['fitnessRelative'] * 1000000000000000;
-  });
-
-  return $fitnessRelative;
-}
-
-/**
- * @param $searched
- * @param $decoded
- * @return float
- * @throws SolutionException
- */
-function calculateFittness($searched, $decoded) {
-  $valid = cleanUp($decoded);
-  # echo "$valid ";
-
-  $result = eval("return $valid;");
-  # echo "= $result >> ";
-
-  $difference = $searched-$result;
-  if ($difference == 0) {
-    throw new \SolutionException($valid);
-  }
-
-  return 1/($difference);
-}
-
-/**
- * @param $code
- * @return mixed|string
- */
-function cleanUp($code) {
-  $numbers = ['0','1','2','3','4','5','6','7','8','9'];
-  $signs = ['+', '-', '*', '/'];
-  $needsSign = false;
-  $result = [];
-
-  foreach (str_split($code) as $diggest) {
-    if ($needsSign === false && in_array($diggest, $numbers)) {
-      $result[] = $diggest;
-      $needsSign = true;
-    }
-
-    if ($needsSign && in_array($diggest, $signs)) {
-      $result[] = $diggest;
-      $needsSign = false;
-    }
-  }
-
-  // make sure last diggest is a number
-  $last = array_pop($result);
-  if (in_array($last, $numbers)) {
-    $result[] = $last;
-  }
-
-  // avoid division by zero errors
-  $returnValue = implode('', $result);
-  $returnValue = str_replace('/0', '', $returnValue);
-
-  if (empty($returnValue)) {
-    return '0';
-  }
-
-  return $returnValue;
-}
-
-/**
- * @param array $population
- * @return array
- */
-function calcRelativeFitness(array $population) {
-  $returnValue = [];
-  $sum = calcFitnessSum($population);
-
-  foreach($population as $values) {
-    $returnValue[] = [
-      'chromosome' => $values['chromosome'],
-      'fitnessAbsolute' => $values[ 'fitnessAbsolute' ],
-      'fitnessRelative' => $values['fitnessAbsolute'] / $sum
-    ];
-  }
-
-  return $returnValue;
-}
-
-/**
- * @param array $population
- * @return int
- */
-function calcFitnessSum(array $population) {
-  $sum = 0;
-  foreach($population as $values) {
-      $sum += $values[ 'fitnessAbsolute' ];
-  }
-  return $sum;
-}
-
-/**
- * @param $population
- * @param ConfigurationValues $c
- * @return array
- * @throws SelectionException
- */
-function selectPair($population, ConfigurationValues $c) {
-  $spez1 = ['chromosome' => null];
-  $spez2 = ['chromosome' => null];
-  $counter = 1;
-
-   while($spez1['chromosome'] == $spez2['chromosome']) {
-    $spez1 = selectSpez($population);
-    $spez2 = selectSpez($population);
-
-    if ($counter >= $c->MaxSelectionAttempts()) {
-      throw new SelectionException($c->MaxSelectionAttempts(), $population);
-    }
-    $counter++;
-  }
-
-  return [ $spez1, $spez2 ];
-}
-
-/**
- * @param array $population keys: ['chromosome', 'fitnessAbsolute', 'fitnessRelative']
- * @expects $population sorted ASC by 'fitnessRelative'
- * @return [$chromosome1, $chromosome2]
- */
-function selectSpez($population) {
-  $sum = calcFitnessSum($population);
-  // rand return 0 or 1 so we have to use integer which can
-  // represent our full precision
-  $randWeight = rand(0,1000000000000000) / 1000000000000000;
-
-  foreach ($population as $index => $value) {
-    $randWeight -= $value['fitnessRelative'];
-    if ($randWeight <= 0) {
-      return $population[$index];
-    }
-  }
-
-  return $population[count($population) - 1];
-}
-
-/**
- * @param $spez1
- * @param $spez2
- * @param ConfigurationValues $c
- * @return array
- */
-function sex($spez1, $spez2, ConfigurationValues $c) {
-  // todo: mann: verify performance if only one child is returned
-  // todo: mann: verify performance if only sometimes multiple childs are returned
-
-  // @todo mann: verify performance if mutation only happens during recombination
-  $random = rand(0,100) / 100;
-  $child1 = '';
-  $child2 = '';
-
-  // recombination
-  if ($random <= $c->CrossoverRate()) {
-    // @todo mann: verify performance if recombinationPoint got new random number
-    $length = array_sum(count_chars($spez1));
-    $recombinationPoint = intval($length * $random);
-    $child1 = substr($spez1, 0 , $recombinationPoint) . substr($spez2, $recombinationPoint, $length);
-    $child2 = substr($spez2, 0 , $recombinationPoint) . substr($spez1, $recombinationPoint, $length);
-  }
-
-  // mutation: without we find the result in the first generation or nevert (local maximum???)
-  $child1 = mutate($child1, $c->MutationRate());
-  $child2 = mutate($child2, $c->MutationRate());
-
-  return [$child1, $child2];
-}
-
-/**
- * @param $chromosome
- * @param $mutationRate
- * @return string
- */
-function mutate($chromosome, $mutationRate) {
-  $bits = str_split($chromosome);
-
-  foreach ($bits as $pos => $bit) {
-    $random = rand(0,1000000) / 1000000;
-    if ($random <= $mutationRate) {
-      $bit = ($bit == '1') ? '0' : '1';
-      $bits[$pos] = $bit;
-    }
-  }
-
-  return implode('', $bits);
-}
-
